Moved tests around, got first sep-6 test working
